# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Cashflow Minimal API",
    "version": "1.0.0",
    "description": "Read-only Lunch Money ingest + local cashflow controls. Minimal endpoints for GPT Actions."
  },
  "servers": [
    { "url": "https://cashflow.mira.joseai.dev" }
  ],
  "paths": {
...
    "/sync": {
      "post": {
        "summary": "Sync LM → DB (insert-only into transactions)",
        "parameters": [
          { "in": "query", "name": "days", "schema": { "type": "integer", "minimum": 1 }, "required": false, "description": "Default 1; use 90 once for initial backfill" },
          { "in": "query", "name": "start", "schema": { "type": "string", "format": "date" }, "required": false },
          { "in": "query", "name": "end", "schema": { "type": "string", "format": "date" }, "required": false }
        ],
        "responses": {
          "200": {
            "description": "Sync summary",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SyncResponse" } } }
          }
        }
      }
    },
    "/transactions": {
      "get": {
        "summary": "List transactions",
        "parameters": [
          { "in": "query", "name": "from", "schema": { "type": "string", "format": "date" } },
          { "in": "query", "name": "to", "schema": { "type": "string", "format": "date" } },
          { "in": "query", "name": "ignored", "schema": { "type": "boolean" } },
          { "in": "query", "name": "limit", "schema": { "type": "integer", "default": 200, "maximum": 1000 } },
          { "in": "query", "name": "offset", "schema": { "type": "integer", "default": 0 } }
        ],
        "responses": {
          "200": {
            "description": "Array of transactions",
            "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Transaction" } } } }
          }
        }
      }
    },
    "/transactions/{lm_id}": {
      "get": {
        "summary": "Get a transaction",
        "parameters": [{ "in": "path", "name": "lm_id", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Transaction", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Transaction" } } } }, "404": { "description": "Not found" } }
      },
      "patch": {
        "summary": "Patch a transaction (category, ignored, note, payee)",
        "parameters": [{ "in": "path", "name": "lm_id", "required": true, "schema": { "type": "integer" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TransactionPatch" } } } },
        "responses": { "200": { "description": "Updated", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Transaction" } } } }, "404": { "description": "Not found" } }
      }
    },
    "/categories": {
      "get": {
        "summary": "List categories",
        "responses": { "200": { "description": "Array", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Category" } } } } } }
      },
      "post": {
        "summary": "Create category",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CategoryCreate" } } } },
        "responses": { "201": { "description": "Created", "content": { "application/json": { "schema": { "type": "object", "properties": { "ok": { "type": "boolean" }, "id": { "type": "string" } }, "required": ["ok","id"] } } } } }
      }
    },
    "/categories/{id}": {
      "patch": {
        "summary": "Patch category flags/name",
        "parameters": [{ "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CategoryPatch" } } } },
        "responses": { "200": { "description": "OK" }, "404": { "description": "Not found" } }
      },
      "delete": {
        "summary": "Delete category",
        "parameters": [{ "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }],
        "responses": { "204": { "description": "Deleted" }, "404": { "description": "Not found" } }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Transaction": {
        "type": "object",
        "properties": {
          "lm_id": { "type": "integer" },
          "date_posted": { "type": "string", "format": "date" },
          "amount": { "type": "number" },
          "currency": { "type": "string" },
          "payee": { "type": "string", "nullable": true },
          "note": { "type": "string", "nullable": true },
          "category": { "type": "string", "nullable": true },
          "ignored": { "type": "boolean" }
        },
        "required": ["lm_id","date_posted","amount","currency","ignored"]
      },
      "TransactionPatch": {
        "type": "object",
        "properties": {
          "category": { "type": "string", "nullable": true },
          "ignored": { "type": "boolean" },
          "note": { "type": "string", "nullable": true },
          "payee": { "type": "string", "nullable": true }
        },
        "additionalProperties": false
      },
      "Category": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "affects_cashflow": { "type": "boolean" },
          "budgetable": { "type": "boolean" },
          "is_income": { "type": "boolean" }
        },
        "required": ["id","name","affects_cashflow","budgetable","is_income"]
      },
      "CategoryCreate": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "affects_cashflow": { "type": "boolean", "default": true },
          "budgetable": { "type": "boolean", "default": true },
          "is_income": { "type": "boolean", "default": false }
        },
        "required": ["id","name"]
      },
      "CategoryPatch": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "affects_cashflow": { "type": "boolean" },
          "budgetable": { "type": "boolean" },
          "is_income": { "type": "boolean" }
        },
        "additionalProperties": false
      },
...
      "SyncResponse": {
        "type": "object",
        "properties": {
          "start": { "type": "string" },
          "end": { "type": "string" },
          "fetched": { "type": "integer" },
          "upserted_into_lm_transactions": { "type": "integer" },
          "inserted_into_transactions": { "type": "integer" },
          "accounts_processed": { "type": "integer" },
          "balance_snapshots": { "type": "integer" }
        }
      },
...
    }
  }
}